+++ firmware/lin-coolant-temperature/ntc_thermistor.h
@@ -5,14 +5,14 @@
 
 class NTCThermistor {
   public:
-    NTCThermistor(int t0, int t1, int r0, int r1, const int *table, int count, int offset);
+    NTCThermistor(int t0, int t1, int r0, int r1, const long int *table, int count, int offset);
     int32_t lookup(int rth);
 #ifdef FLOATING_POINT_MATH
     int32_t calculate(int rth);
 #endif
 
   protected:
-    const int *_table;
+    const long int *_table;
     int _count;
     int _offset;
